optimize the framework of realtime communication
