+++ src/devices/arduino/arduino-peripheral.js
@@ -1,0 +1,474 @@
+const formatMessage = require('format-message');
+const Buffer = require('buffer').Buffer;
+
+const Serialport = require('../../io/serialport');
+const Base64Util = require('../../util/base64-util');
+
+const Firmata = require('../firmata/firmata');
+
+/**
+ * A string to report connect firmata timeout.
+ * @type {formatMessage}
+ */
+const ConnectFirmataTimeout = formatMessage({
+    id: 'arduinoPeripheral.connection.connectFirmataTimeout',
+    default: 'Timeout when try to connect firmata, please download the firmware first',
+    description: 'label for connect firmata timeout'
+});
+
+/**
+ * A time interval to send firmata heartbeat(in milliseconds).
+ */
+const FrimataHeartbeatInterval = 2000;
+
+/**
+ * A time interval to wait (in milliseconds) before reporting to the serialport socket
+ * that heartbeat has stopped coming from the peripheral.
+ */
+const FrimataHeartbeatTimeout = 5000;
+
+const Level = {
+    High: 'HIGH',
+    Low: 'LOW'
+};
+
+const Mode = {
+    Input: 'INPUT',
+    Output: 'OUTPUT',
+    InputPullup: 'INPUT_PULLUP'
+};
+
+/**
+ * Manage communication with a Arduino peripheral over a OpenBlock Link client socket.
+ */
+class ArduinoPeripheral{
+
+    /**
+     * Construct a Arduino communication object.
+     * @param {Runtime} runtime - the OpenBlock runtime
+     * @param {string} deviceId - the id of the peripheral
+     * @param {object} pnpidList - the pnp id of the peripheral
+     * @param {object} serialConfig - the serial config of the peripheral
+     * @param {object} diveceOpt - the device optione of the peripheral
+     */
+    constructor (runtime, deviceId, pnpidList, serialConfig, diveceOpt) {
+        /**
+         * The OpenBlock runtime used to trigger the green flag button.
+         * @type {Runtime}
+         * @private
+         */
+        this._runtime = runtime;
+
+        this.pnpidList = pnpidList;
+        this.serialConfig = serialConfig;
+        this.diveceOpt = diveceOpt;
+
+        /**
+         * The serialport connection socket for reading/writing peripheral data.
+         * @type {SERIALPORT}
+         * @private
+         */
+        this._serialport = null;
+        this._runtime.registerPeripheralExtension(deviceId, this);
+        this._runtime.setRealtimeBaudrate(this.serialConfig.baudRate);
+
+        /**
+         * The id of the peripheral this peripheral belongs to.
+         */
+        this._deviceId = deviceId;
+
+        /**
+        * Pending data list. If busy is set when send, the data will push into this array to
+        * waitting to be sended.
+        */
+        this._pendingData = [];
+
+        this.reset = this.reset.bind(this);
+        this._onConnect = this._onConnect.bind(this);
+        this._onMessage = this._onMessage.bind(this);
+
+        /**
+         * Firmata connection.
+         * @type {?Firmata}
+         * @private
+         */
+        this._firmata = null;
+
+        /**
+         * Timeout ID for firmata get heartbeat timeout.
+         * @type {number}
+         * @private
+         */
+        this._firmataTimeoutID = null;
+
+        /**
+         * Interval ID for firmata send heartbeat.
+         * @type {number}
+         * @private
+         */
+        this._firmataIntervelID = null;
+
+        /**
+         * A flag that is true while firmata is conncted.
+         * @type {boolean}
+         * @private
+         */
+        this._isFirmataConnected = false;
+
+        this._startHeartbeat = this._startHeartbeat.bind(this);
+        this._listenHeartbeat = this._listenHeartbeat.bind(this);
+        this._handleProgramModeUpdate = this._handleProgramModeUpdate.bind(this);
+    }
+
+    /**
+     * Called by the runtime when user wants to upload code to a peripheral.
+     * @param {string} code - the code want to upload.
+     */
+    upload (code) {
+        const base64Str = Buffer.from(code).toString('base64');
+        this._serialport.upload(base64Str, this.diveceOpt, 'base64');
+    }
+
+    /**
+     * Called by the runtime when user wants to upload realtime firmware to a peripheral.
+     */
+    uploadFirmware () {
+        this._stopHeartbeat();
+        this._serialport.uploadFirmware(this.diveceOpt);
+    }
+
+    /**
+     * Called by the runtime when user wants to scan for a peripheral.
+     * @param {Array.<string>} pnpidList - the array of pnp id list
+     * @param {bool} listAll - wether list all connectable device
+     */
+    scan (pnpidList, listAll) {
+        if (this._serialport) {
+            this._serialport.disconnect();
+        }
+        this._serialport = new Serialport(this._runtime, this._deviceId, {
+            filters: {
+                pnpid: listAll ? ['*'] : (pnpidList ? pnpidList : this.pnpidList)
+            }
+        }, this._onConnect, this.reset);
+    }
+
+    /**
+     * Called by the runtime when user wants to connect to a certain peripheral.
+     * @param {number} id - the id of the peripheral to connect to.
+     * @param {?number} baudrate - the baudrate.
+     */
+    connect (id, baudrate = null) {
+        const config = this.serialConfig;
+        if (baudrate) {
+            config.baudRate = baudrate;
+        }
+        if (this._serialport) {
+            this._serialport.connectPeripheral(id, {config: config});
+        }
+    }
+
+    /**
+     * Disconnect from the peripheral.
+     */
+    disconnect () {
+        if (this._serialport) {
+            this._serialport.disconnect();
+        }
+
+        this.reset();
+    }
+
+    /**
+     * Reset all the state and timeout/interval ids.
+     */
+    reset () {
+        if (this._firmata) {
+            this._firmata.removeListener('reportversion', this._listenHeartbeat);
+            delete this._firmata;
+        }
+        this._stopHeartbeat();
+        this._runtime.removeListener(this._runtime.constructor.PROGRAM_MODE_UPDATE, this._handleProgramModeUpdate);
+        this._runtime.removeListener(this._runtime.constructor.PERIPHERAL_UPLOAD_SUCCESS, this._startHeartbeat);
+
+        this._isFirmataConnected = false;
+    }
+
+    /**
+     * Return true if connected to the peripheral.
+     * @return {boolean} - whether the peripheral is connected.
+     */
+    isConnected () {
+        let connected = false;
+        if (this._serialport) {
+            connected = this._serialport.isConnected();
+        }
+        return connected;
+    }
+
+    /**
+     * Set baudrate of the peripheral serialport.
+     * @param {number} baudrate - the baudrate.
+     */
+    setBaudrate (baudrate) {
+        this._serialport.setBaudrate(baudrate);
+    }
+
+    /**
+     * Write data to the peripheral serialport.
+     * @param {string} data - the data to write.
+     */
+    write (data) {
+        if (!this.isConnected()) return;
+
+        const base64Str = data.toString('base64');
+        this._serialport.write(base64Str, 'base64');
+    }
+
+    /**
+     * Send a message to the peripheral Serialport socket.
+     * @param {Uint8Array} message - the message to write
+     */
+    send (message) {
+        if (!this.isConnected()) return;
+
+        const data = Base64Util.uint8ArrayToBase64(message);
+        this._serialport.write(data, 'base64');
+    }
+
+    /**
+     * Start send/recive heartbeat timer.
+     * @private
+     */
+    _startHeartbeat () {
+        if (this._runtime.getCurrentIsRealtimeMode()) {
+            this._isFirmataConnected = false;
+
+            this._firmataIntervelID = window.setInterval(() => {
+                if (this._runtime.getCurrentIsRealtimeMode()) {
+                // Send reportVersion request as heartbeat.
+                    this._firmata.reportVersion(() => { });
+                }
+            }, FrimataHeartbeatInterval);
+            // Start a timer if heartbeat time out means failed to connect firmata.
+            this._firmataTimeoutID = window.setTimeout(() => {
+                this._isFirmataConnected = false;
+                this._serialport.handleRealtimeDisconnectError(ConnectFirmataTimeout);
+            }, FrimataHeartbeatTimeout);
+        }
+    }
+
+    /**
+     * Stop send/recive heartbeat timer.
+     * @private
+     */
+    _stopHeartbeat () {
+        if (this._firmataTimeoutID) {
+            window.clearTimeout(this._firmataTimeoutID);
+            this._firmataTimeoutID = null;
+        }
+        if (this._firmataIntervelID) {
+            window.clearInterval(this._firmataIntervelID);
+            this._firmataIntervelID = null;
+        }
+        this._isFirmataConnected = false;
+    }
+
+    /**
+     * Listen the heartbeat and emit connection state event.
+     * @private
+     */
+    _listenHeartbeat () {
+        if (!this._isFirmataConnected) {
+            this._isFirmataConnected = true;
+            if (this._runtime.getCurrentIsRealtimeMode()) {
+                this._serialport.handleRealtimeConnectSucess();
+            }
+        }
+        // Reset the timeout timer
+        window.clearTimeout(this._firmataTimeoutID);
+        this._firmataTimeoutID = window.setTimeout(() => {
+            this._isFirmataConnected = false;
+            this._serialport.handleRealtimeDisconnectError(ConnectFirmataTimeout);
+        }, FrimataHeartbeatTimeout);
+    }
+
+    /**
+     * Handle the program mode update event. If in realtime mode start the heartbeat else stop.
+     */
+    _handleProgramModeUpdate () {
+        if (this._runtime.getCurrentIsRealtimeMode()) {
+            this._startHeartbeat();
+        } else {
+            this._stopHeartbeat();
+        }
+    }
+
+    /**
+     * Starts reading data from peripheral after serialport has connected to it.
+     * @private
+     */
+    _onConnect () {
+        this._serialport.read(this._onMessage);
+        this._firmata = new Firmata(this.send.bind(this));
+
+        this._startHeartbeat();
+
+        this._runtime.on(this._runtime.constructor.PROGRAM_MODE_UPDATE, this._handleProgramModeUpdate);
+        this._runtime.on(this._runtime.constructor.PERIPHERAL_UPLOAD_SUCCESS, this._startHeartbeat);
+        // Start the heartbeat listener.
+        this._firmata.on('reportversion', this._listenHeartbeat);
+    }
+
+    /**
+     * Process the sensor data from the incoming serialport characteristic.
+     * @param {object} base64 - the incoming serialport data.
+     * @private
+     */
+    _onMessage (base64) {
+        // parse data
+        const data = Base64Util.base64ToUint8Array(base64);
+        this._firmata.onReciveData(data);
+    }
+
+    /**
+     * Return true if peripheral has connected to firmata and program mode is realtime.
+     * @return {boolean} - whether the peripheral is ready for realtime mode communication.
+     */
+    isReady () {
+        if (this._runtime.getCurrentIsRealtimeMode() && this._isFirmataConnected) {
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * @param {PIN} pin - the pin string to parse.
+     * @return {number} - the pin number.
+     */
+    parsePin (pin) {
+        if (pin.charAt(0) === 'A') {
+            return parseInt(pin.slice(1), 10) + 14;
+        }
+        return parseInt(pin, 10);
+    }
+
+    /**
+     * @param {LEVEL} level - the level string to parse.
+     * @return {number} - the level in number.
+     */
+    parseLevel (level) {
+        if (level === Level.High) {
+            return 1;
+        }
+        return 0;
+    }
+
+    /**
+     * @param {PIN} pin - the pin to set.
+     * @param {MODE} mode - the pin mode to set.
+     */
+    setPinMode (pin, mode) {
+        if (this.isReady()) {
+            pin = this.parsePin(pin);
+            switch (mode) {
+            case Mode.Input:
+                mode = this._firmata.MODES.INPUT;
+                break;
+            case Mode.Output:
+                mode = this._firmata.MODES.OUTPUT;
+                break;
+            case Mode.InputPullup:
+                mode = this._firmata.MODES.PULLUP;
+                break;
+            }
+            this._firmata.pinMode(pin, mode);
+        }
+    }
+
+    /**
+     * @param {PIN} pin - the pin to set.
+     * @param {LEVEL} level - the pin level to set.
+     */
+    setDigitalOutput (pin, level) {
+        if (this.isReady()) {
+            pin = this.parsePin(pin);
+            level = this.parseLevel(level);
+            this._firmata.digitalWrite(pin, level);
+        }
+    }
+
+    /**
+     * @param {PIN} pin - the pin to set.
+     * @param {VALUE} value - the pwm value to set.
+     */
+    setPwmOutput (pin, value) {
+        if (this.isReady()) {
+            pin = this.parsePin(pin);
+            if (value < 0) {
+                value = 0;
+            }
+            if (value > 255) {
+                value = 255;
+            }
+            this._firmata.pinMode(pin, this._firmata.MODES.PWM);
+            this._firmata.pwmWrite(pin, value);
+        }
+    }
+
+    /**
+     * @param {PIN} pin - the pin to read.
+     * @return {Promise} - a Promise that resolves when read from peripheral.
+     */
+    readDigitalPin (pin) {
+        if (this.isReady()) {
+            pin = this.parsePin(pin);
+            return new Promise(resolve => {
+                this._firmata.digitalRead(pin, value => {
+                    resolve(value);
+                });
+            });
+        }
+    }
+
+    /**
+     * @param {PIN} pin - the pin to read.
+     * @return {Promise} - a Promise that resolves when read from peripheral.
+     */
+    readAnalogPin (pin) {
+        if (this.isReady()) {
+            pin = this.parsePin(pin);
+            // Shifting to analog pin number.
+            pin = pin - 14;
+            this._firmata.pinMode(pin, this._firmata.MODES.ANALOG);
+            return new Promise(resolve => {
+                this._firmata.analogRead(pin, value => {
+                    resolve(value);
+                });
+            });
+        }
+    }
+
+    /**
+     * @param {PIN} pin - the pin to set.
+     * @param {VALUE} value - the degree to set.
+     */
+    setServoOutput (pin, value) {
+        if (this.isReady()) {
+            pin = this.parsePin(pin);
+            if (value < 0) {
+                value = 0;
+            }
+            if (value > 180) {
+                value = 180;
+            }
+            this._firmata.pinMode(pin, this._firmata.MODES.PWM);
+            this._firmata.pwmWrite(pin, value);
+
+            this._firmata.servoConfig(pin, 600, 2400);
+            this._firmata.servoWrite(pin, value);
+        }
+    }
+}
+
+module.exports = ArduinoPeripheral;
