+++ src/devices/arduinoMega2560/index.js
@@ -1,13 +1,10 @@
 const formatMessage = require('format-message');
-const Buffer = require('buffer').Buffer;
 
 const ArgumentType = require('../../extension-support/argument-type');
 const BlockType = require('../../extension-support/block-type');
 const ProgramModeType = require('../../extension-support/program-mode-type');
-const Serialport = require('../../io/serialport');
-const Base64Util = require('../../util/base64-util');
 
-const Firmata = require('../../lib/firmata/firmata');
+const ArduinoPeripheral = require('../arduino/arduino-peripheral');
 
 /**
  * The list of USB device filters.
@@ -44,27 +41,6 @@ const DIVECE_OPT = {
     fqbn: 'arduino:avr:mega:cpu=atmega2560',
     firmware: 'arduinoMega2560.standardFirmata.ino.hex'
 };
-
-/**
- * A string to report connect firmata timeout.
- * @type {formatMessage}
- */
-const ConnectFirmataTimeout = formatMessage({
-    id: 'arduinoMega2560.connection.connectFirmataTimeout',
-    default: 'Timeout when try to connect firmata, please download the firmware first',
-    description: 'label for connect firmata timeout'
-});
-
-/**
- * A time interval to send firmata heartbeat(in milliseconds).
- */
-const FrimataHeartbeatInterval = 2000;
-
-/**
- * A time interval to wait (in milliseconds) before reporting to the serialport socket
- * that heartbeat has stopped coming from the peripheral.
- */
-const FrimataHeartbeatTimeout = 5000;
 
 const Pins = {
     D0: '0',
@@ -188,420 +164,14 @@ const DataType = {
 /**
  * Manage communication with a Arduino Mega2560 peripheral over a OpenBlock Link client socket.
  */
-class ArduinoMega2560{
-
+class ArduinoMega2560 extends ArduinoPeripheral{
     /**
      * Construct a Arduino communication object.
      * @param {Runtime} runtime - the OpenBlock runtime
      * @param {string} deviceId - the id of the extension
      */
     constructor (runtime, deviceId) {
-        /**
-         * The OpenBlock runtime used to trigger the green flag button.
-         * @type {Runtime}
-         * @private
-         */
-        this._runtime = runtime;
-
-        /**
-         * The serialport connection socket for reading/writing peripheral data.
-         * @type {SERIALPORT}
-         * @private
-         */
-        this._serialport = null;
-        this._runtime.registerPeripheralExtension(deviceId, this);
-        this._runtime.setRealtimeBaudrate(SERIAL_CONFIG.baudRate);
-
-        /**
-         * The id of the extension this peripheral belongs to.
-         */
-        this._deviceId = deviceId;
-
-        /**
-        * Pending data list. If busy is set when send, the data will push into this array to
-        * waitting to be sended.
-        */
-        this._pendingData = [];
-
-        this.reset = this.reset.bind(this);
-        this._onConnect = this._onConnect.bind(this);
-        this._onMessage = this._onMessage.bind(this);
-
-        /**
-         * Firmata connection.
-         * @type {?Firmata}
-         * @private
-         */
-        this._firmata = null;
-
-        /**
-         * Timeout ID for firmata get heartbeat timeout.
-         * @type {number}
-         * @private
-         */
-        this._firmataTimeoutID = null;
-
-        /**
-         * Interval ID for firmata send heartbeat.
-         * @type {number}
-         * @private
-         */
-        this._firmataIntervelID = null;
-
-        /**
-         * A flag that is true while firmata is conncted.
-         * @type {boolean}
-         * @private
-         */
-        this._isFirmataConnected = false;
-
-        this._startHeartbeat = this._startHeartbeat.bind(this);
-        this._listenHeartbeat = this._listenHeartbeat.bind(this);
-    }
-
-    /**
-     * Called by the runtime when user wants to upload code to a peripheral.
-     * @param {string} code - the code want to upload.
-     */
-    upload (code) {
-        const base64Str = Buffer.from(code).toString('base64');
-        this._serialport.upload(base64Str, DIVECE_OPT, 'base64');
-    }
-
-    /**
-     * Called by the runtime when user wants to upload realtime firmware to a peripheral.
-     */
-    uploadFirmware () {
-        this._stopHeartbeat();
-        this._serialport.uploadFirmware(DIVECE_OPT);
-    }
-
-    /**
-     * Called by the runtime when user wants to scan for a peripheral.
-     * @param {Array.<string>} pnpidList - the array of pnp id list
-     * @param {bool} listAll - wether list all connectable device
-     */
-    scan (pnpidList, listAll) {
-        if (this._serialport) {
-            this._serialport.disconnect();
-        }
-        this._serialport = new Serialport(this._runtime, this._deviceId, {
-            filters: {
-                pnpid: listAll ? ['*'] : (pnpidList ? pnpidList : PNPID_LIST)
-            }
-        }, this._onConnect, this.reset);
-    }
-
-    /**
-     * Called by the runtime when user wants to connect to a certain peripheral.
-     * @param {number} id - the id of the peripheral to connect to.
-     * @param {?number} baudrate - the baudrate.
-     */
-    connect (id, baudrate = null) {
-        const config = SERIAL_CONFIG;
-        if (baudrate) {
-            config.baudRate = baudrate;
-        }
-        if (this._serialport) {
-            this._serialport.connectPeripheral(id, {config: config});
-        }
-    }
-
-    /**
-     * Disconnect from the peripheral.
-     */
-    disconnect () {
-        if (this._serialport) {
-            this._serialport.disconnect();
-        }
-
-        this.reset();
-    }
-
-    /**
-     * Reset all the state and timeout/interval ids.
-     */
-    reset () {
-        if (this._firmata) {
-            this._firmata.removeListener('reportversion', this._listenHeartbeat);
-            delete this._firmata;
-        }
-        this._stopHeartbeat();
-        this._runtime.removeListener(this._runtime.constructor.PROGRAM_MODE_UPDATE, this._startHeartbeat);
-        this._runtime.removeListener(this._runtime.constructor.PERIPHERAL_UPLOAD_SUCCESS, this._startHeartbeat);
-
-        this._isFirmataConnected = false;
-    }
-
-    /**
-     * Return true if connected to the peripheral.
-     * @return {boolean} - whether the peripheral is connected.
-     */
-    isConnected () {
-        let connected = false;
-        if (this._serialport) {
-            connected = this._serialport.isConnected();
-        }
-        return connected;
-    }
-
-    /**
-     * Set baudrate of the peripheral serialport.
-     * @param {number} baudrate - the baudrate.
-     */
-    setBaudrate (baudrate) {
-        this._serialport.setBaudrate(baudrate);
-    }
-
-    /**
-     * Write data to the peripheral serialport.
-     * @param {string} data - the data to write.
-     */
-    write (data) {
-        if (!this.isConnected()) return;
-
-        const base64Str = data.toString('base64');
-        this._serialport.write(base64Str, 'base64');
-    }
-
-    /**
-     * Send a message to the peripheral Serialport socket.
-     * @param {Uint8Array} message - the message to write
-     */
-    send (message) {
-        if (!this.isConnected()) return;
-
-        const data = Base64Util.uint8ArrayToBase64(message);
-        this._serialport.write(data, 'base64');
-    }
-
-    /**
-     * Start send/recive heartbeat timer.
-     * @private
-     */
-    _startHeartbeat () {
-        if (this._runtime.getCurrentIsRealtimeMode()) {
-            this._isFirmataConnected = false;
-
-            this._firmataIntervelID = window.setInterval(() => {
-                if (this._runtime.getCurrentIsRealtimeMode()) {
-                // Send reportVersion request as heartbeat.
-                    this._firmata.reportVersion(() => { });
-                }
-            }, FrimataHeartbeatInterval);
-            // Start a timer if heartbeat time out means failed to connect firmata.
-            this._firmataTimeoutID = window.setTimeout(() => {
-                this._isFirmataConnected = false;
-                if (this._runtime.getCurrentIsRealtimeMode()) {
-                    this._serialport.handleRealtimeDisconnectError(ConnectFirmataTimeout);
-                }
-            }, FrimataHeartbeatTimeout);
-        }
-    }
-
-    /**
-     * Stop send/recive heartbeat timer.
-     * @private
-     */
-    _stopHeartbeat () {
-        if (this._firmataTimeoutID) {
-            window.clearTimeout(this._firmataTimeoutID);
-            this._firmataTimeoutID = null;
-        }
-        if (this._firmataIntervelID) {
-            window.clearInterval(this._firmataIntervelID);
-            this._firmataIntervelID = null;
-        }
-        this._isFirmataConnected = false;
-    }
-
-    /**
-     * Listen the heartbeat and emit connection state event.
-     * @private
-     */
-    _listenHeartbeat () {
-        if (!this._isFirmataConnected) {
-            this._isFirmataConnected = true;
-            if (this._runtime.getCurrentIsRealtimeMode()) {
-                this._serialport.handleRealtimeConnectSucess();
-            }
-        }
-        // Reset the timeout timer
-        window.clearTimeout(this._firmataTimeoutID);
-        this._firmataTimeoutID = window.setTimeout(() => {
-            this._isFirmataConnected = false;
-            if (this._runtime.getCurrentIsRealtimeMode()) {
-                this._serialport.handleRealtimeDisconnectError(ConnectFirmataTimeout);
-            }
-        }, FrimataHeartbeatTimeout);
-    }
-
-    /**
-     * Starts reading data from peripheral after serialport has connected to it.
-     * @private
-     */
-    _onConnect () {
-        this._serialport.read(this._onMessage);
-        this._firmata = new Firmata(this.send.bind(this));
-
-        if (this._runtime.getCurrentIsRealtimeMode()) {
-            this._startHeartbeat();
-        }
-
-        this._runtime.on(this._runtime.constructor.PROGRAM_MODE_UPDATE, this._startHeartbeat);
-        this._runtime.on(this._runtime.constructor.PERIPHERAL_UPLOAD_SUCCESS, this._startHeartbeat);
-        // Start the heartbeat listener.
-        this._firmata.on('reportversion', this._listenHeartbeat);
-    }
-
-    /**
-     * Process the sensor data from the incoming serialport characteristic.
-     * @param {object} base64 - the incoming serialport data.
-     * @private
-     */
-    _onMessage (base64) {
-        // parse data
-        const data = Base64Util.base64ToUint8Array(base64);
-        this._firmata.onReciveData(data);
-    }
-
-    /**
-     * Return true if peripheral has connected to firmata and program mode is realtime.
-     * @return {boolean} - whether the peripheral is ready for realtime mode communication.
-     */
-    isReady () {
-        if (this._runtime.getCurrentIsRealtimeMode() && this._isFirmataConnected) {
-            return true;
-        }
-        return false;
-    }
-
-    /**
-     * @param {PIN} pin - the pin string to parse.
-     * @return {number} - the pin number.
-     */
-    parsePin (pin) {
-        if (pin.charAt(0) === 'A') {
-            return parseInt(pin.slice(1), 10) + 14;
-        }
-        return parseInt(pin, 10);
-    }
-
-    /**
-     * @param {LEVEL} level - the level string to parse.
-     * @return {number} - the level in number.
-     */
-    parseLevel (level) {
-        if (level === Level.High) {
-            return 1;
-        }
-        return 0;
-    }
-
-    /**
-     * @param {PIN} pin - the pin to set.
-     * @param {MODE} mode - the pin mode to set.
-     */
-    setPinMode (pin, mode) {
-        if (this.isReady()) {
-            pin = this.parsePin(pin);
-            switch (mode) {
-            case Mode.Input:
-                mode = this._firmata.MODES.INPUT;
-                break;
-            case Mode.Output:
-                mode = this._firmata.MODES.OUTPUT;
-                break;
-            case Mode.InputPullup:
-                mode = this._firmata.MODES.PULLUP;
-                break;
-            }
-            this._firmata.pinMode(pin, mode);
-        }
-    }
-
-    /**
-     * @param {PIN} pin - the pin to set.
-     * @param {LEVEL} level - the pin level to set.
-     */
-    setDigitalOutput (pin, level) {
-        if (this.isReady()) {
-            pin = this.parsePin(pin);
-            level = this.parseLevel(level);
-            this._firmata.digitalWrite(pin, level);
-        }
-    }
-
-    /**
-     * @param {PIN} pin - the pin to set.
-     * @param {VALUE} value - the pwm value to set.
-     */
-    setPwmOutput (pin, value) {
-        if (this.isReady()) {
-            pin = this.parsePin(pin);
-            if (value < 0) {
-                value = 0;
-            }
-            if (value > 255) {
-                value = 255;
-            }
-            this._firmata.pinMode(pin, this._firmata.MODES.PWM);
-            this._firmata.pwmWrite(pin, value);
-        }
-    }
-
-    /**
-     * @param {PIN} pin - the pin to read.
-     * @return {Promise} - a Promise that resolves when read from peripheral.
-     */
-    readDigitalPin (pin) {
-        if (this.isReady()) {
-            pin = this.parsePin(pin);
-            return new Promise(resolve => {
-                this._firmata.digitalRead(pin, value => {
-                    resolve(value);
-                });
-            });
-        }
-    }
-
-    /**
-     * @param {PIN} pin - the pin to read.
-     * @return {Promise} - a Promise that resolves when read from peripheral.
-     */
-    readAnalogPin (pin) {
-        if (this.isReady()) {
-            pin = this.parsePin(pin);
-            // Shifting to analog pin number.
-            pin = pin - 14;
-            this._firmata.pinMode(pin, this._firmata.MODES.ANALOG);
-            return new Promise(resolve => {
-                this._firmata.analogRead(pin, value => {
-                    resolve(value);
-                });
-            });
-        }
-    }
-
-    /**
-     * @param {PIN} pin - the pin to set.
-     * @param {VALUE} value - the degree to set.
-     */
-    setServoOutput (pin, value) {
-        if (this.isReady()) {
-            pin = this.parsePin(pin);
-            if (value < 0) {
-                value = 0;
-            }
-            if (value > 180) {
-                value = 180;
-            }
-            this._firmata.pinMode(pin, this._firmata.MODES.PWM);
-            this._firmata.pwmWrite(pin, value);
-
-            this._firmata.servoConfig(pin, 600, 2400);
-            this._firmata.servoWrite(pin, value);
-        }
+        super(runtime, deviceId, PNPID_LIST, SERIAL_CONFIG, DIVECE_OPT);
     }
 }
